Code cleanups
