+++ src/i18n/en.json
@@ -73,7 +73,8 @@
     "UN_ASSIGN": "Unassign conversation",
     "MUTE_CONVERSATION": "Mute conversation",
     "UNMUTE_CONVERSATION": "Unmute conversation",
-    "CONVERSATION_NOT_FOUND": "Could not fetch conversation"
+    "CONVERSATION_NOT_FOUND": "Could not fetch conversation",
+    "MESSAGES": "Messages"
   },
   "FILTER": {
     "ALL_INBOXES": "All inboxes",
