+++ src/screens/ChatScreen/components/MessageList/MessageList.style.js
@@ -1,0 +1,22 @@
+const styles = theme => ({
+  container: {
+    flex: 1,
+    backgroundColor: theme['color-background'],
+  },
+  chatView: {
+    flex: 13,
+  },
+
+  chatContainer: {
+    paddingHorizontal: 16,
+    paddingVertical: 12,
+    paddingTop: 8,
+  },
+  loadMoreSpinnerView: {
+    alignItems: 'center',
+    justifyContent: 'center',
+    paddingBottom: 8,
+  },
+});
+
+export default styles;
